feat: block author/category deletion with associated books

+++ src/services/author.service.js
@@ -1,5 +1,5 @@
 const { status: httpStatus } = require('http-status');
-const { Author } = require('../models');
+const { Author, Book } = require('../models');
 const ApiError = require('../utils/ApiError');
 
 const createAuthor = async (AuthorBody) => Author.create(AuthorBody);
@@ -11,7 +11,7 @@ const getAuthorById = async (id) => Author.findById(id);
 const updateAuthorById = async (id, body) => {
   const author = await getAuthorById(id);
   if (!author) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'author not found');
+    throw new ApiError(httpStatus.NOT_FOUND, 'Author not found');
   }
   Object.assign(author, body);
   await author.save();
@@ -21,8 +21,14 @@ const updateAuthorById = async (id, body) => {
 const deleteAuthorById = async (id) => {
   const author = await getAuthorById(id);
   if (!author) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'author not found');
+    throw new ApiError(httpStatus.NOT_FOUND, 'Author not found');
   }
+
+  const associatedBooks = await Book.exists({ author: id });
+  if (associatedBooks) {
+    throw new ApiError(httpStatus.CONFLICT, 'Cannot delete author with associated books');
+  }
+
   await author.deleteOne();
   return author;
 };
